Remove sync between surveyjs and form
