+++ apps/frontend/src/pages/Surveys/Editor/SurveyEditorPage.tsx
@@ -10,49 +10,59 @@
  * You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import React, { useEffect, useMemo } from 'react';
+import React, { useEffect, useMemo, useRef } from 'react';
 import { toast } from 'sonner';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
-import { useTranslation } from 'react-i18next';
 import { zodResolver } from '@hookform/resolvers/zod';
-import getInitialSurveyFormValues from '@libs/survey/constants/initial-survey-form';
+import { useTranslation } from 'react-i18next';
+import { settings } from 'survey-core';
+import { editorLocalization, localization } from 'survey-creator-core';
+import { SurveyCreator, SurveyCreatorComponent } from 'survey-creator-react';
+import 'survey-creator-core/i18n/english';
+import 'survey-creator-core/i18n/german';
 import SurveyDto from '@libs/survey/types/api/survey.dto';
 import AttendeeDto from '@libs/user/types/attendee.dto';
+import SurveyFormula from '@libs/survey/types/TSurveyFormula';
+import getInitialSurveyFormValues from '@libs/survey/constants/initial-survey-form';
+import { CREATED_SURVEYS_PAGE } from '@libs/survey/constants/surveys-endpoint';
+import getSurveyEditorFormSchema from '@libs/survey/types/editor/surveyEditorForm.schema';
 import useUserStore from '@/store/UserStore/UserStore';
 import useSurveyTablesPageStore from '@/pages/Surveys/Tables/useSurveysTablesPageStore';
 import useSurveyEditorPageStore from '@/pages/Surveys/Editor/useSurveyEditorPageStore';
-import { CREATED_SURVEYS_PAGE } from '@libs/survey/constants/surveys-endpoint';
-import getSurveyEditorFormSchema from '@libs/survey/types/editor/surveyEditorForm.schema';
-import SurveyEditor from '@/pages/Surveys/Editor/components/SurveyEditor';
-import SaveSurveyDialog from '@/pages/Surveys/Editor/dialog/SaveSurveyDialog';
+import useLanguage from '@/hooks/useLanguage';
+import useElementHeight from '@/hooks/useElementHeight';
 import SaveButton from '@/components/shared/FloatingsButtonsBar/CommonButtonConfigs/saveButton';
 import FloatingButtonsBarConfig from '@libs/ui/types/FloatingButtons/floatingButtonsBarConfig';
 import FloatingButtonsBar from '@/components/shared/FloatingsButtonsBar/FloatingButtonsBar';
 import LoadingIndicatorDialog from '@/components/ui/Loading/LoadingIndicatorDialog';
+import SaveSurveyDialog from '@/pages/Surveys/Editor/dialog/SaveSurveyDialog';
+import surveyTheme from '@/pages/Surveys/theme/theme';
+import '@/pages/Surveys/theme/default2.min.css';
+import '@/pages/Surveys/theme/creator.min.css';
+import '@/pages/Surveys/theme/custom.survey.css';
+import '@/pages/Surveys/theme/custom.creator.css';
+import { FLOATING_BUTTONS_BAR_ID, FOOTER_ID } from '@libs/common/constants/pageElementIds';
+
+settings.lazyRender.enabled = true;
 
 const SurveyEditorPage = () => {
-  const { updateSelectedSurvey, isFetching, selectedSurvey, updateUsersSurveys } = useSurveyTablesPageStore();
+  const { fetchSelectedSurvey, isFetching, selectedSurvey, updateUsersSurveys } = useSurveyTablesPageStore();
   const { isOpenSaveSurveyDialog, setIsOpenSaveSurveyDialog, updateOrCreateSurvey, isLoading, reset } =
     useSurveyEditorPageStore();
-
   const { t } = useTranslation();
   const navigate = useNavigate();
-
   const { user } = useUserStore();
-
   const { surveyId } = useParams();
+  const { language } = useLanguage();
 
   useEffect(() => {
     reset();
-    void updateSelectedSurvey(surveyId, false);
+    void fetchSelectedSurvey(surveyId, false);
   }, [surveyId]);
 
   const initialFormValues: SurveyDto | undefined = useMemo(() => {
-    if (!user || !user.username) {
-      return undefined;
-    }
-
+    if (!user || !user.username) return undefined;
     const surveyCreator: AttendeeDto = {
       firstName: user.firstName,
       lastName: user.lastName,
@@ -73,49 +83,106 @@ const SurveyEditorPage = () => {
     form.reset(initialFormValues);
   }, [initialFormValues]);
 
-  const saveSurvey = async () => {
-    const {
-      id,
-      formula,
-      saveNo,
-      creator,
-      invitedAttendees,
-      invitedGroups,
-      participatedAttendees,
-      answers,
-      createdAt,
-      expires,
-      isAnonymous,
-      isPublic,
-      canSubmitMultipleAnswers,
-      canUpdateFormerAnswer,
-    } = form.getValues();
+  editorLocalization.defaultLocale = language;
+  localization.currentLocale = language;
 
-    const success = await updateOrCreateSurvey({
-      id,
-      formula,
-      saveNo,
-      creator,
-      invitedAttendees,
-      invitedGroups,
-      participatedAttendees,
-      answers,
-      createdAt,
-      expires,
-      isAnonymous,
-      isPublic,
-      canSubmitMultipleAnswers,
-      canUpdateFormerAnswer,
+  const creatorOptions = useMemo(
+    () => ({
+      generateValidJSON: true,
+      isAutoSave: false,
+      maxNestedPanels: 0,
+      showJSONEditorTab: true,
+      showPreviewTab: false,
+      showLogicTab: true,
+      questionTypes: [
+        'radiogroup',
+        'rating',
+        'checkbox',
+        'dropdown',
+        'boolean',
+        'file',
+        'imagepicker',
+        'ranking',
+        'text',
+        'comment',
+        'multipletext',
+        'panel',
+        'paneldynamic',
+        'matrix',
+        'matrixdropdown',
+        'image',
+      ],
+    }),
+    [],
+  );
+
+  // Store the SurveyCreator in a ref so it is only created once
+  const creatorRef = useRef<SurveyCreator | null>(null);
+
+  // Initialize creator only once
+  if (!creatorRef.current) {
+    const creator = new SurveyCreator(creatorOptions);
+    creator.theme = surveyTheme;
+    creator.locale = language;
+    creator.showToolbox = false;
+    creator.showSidebar = false;
+    creator.startEditTitleOnQuestionAdded = true;
+
+    const settingsActionHeader = creator.toolbar.actions.findIndex((action) => action.id === 'svd-settings');
+    if (settingsActionHeader >= 0) creator.toolbar.actions.splice(settingsActionHeader, 1);
+    const expandGridActionHeader = creator.toolbar.actions.findIndex((action) => action.id === 'svd-grid-expand');
+    if (expandGridActionHeader >= 0) creator.toolbar.actions.splice(expandGridActionHeader, 1);
+    const designerActionFooter = creator.footerToolbar.actions.findIndex((action) => action.id === 'svd-designer');
+    if (designerActionFooter >= 0) creator.footerToolbar.actions.splice(designerActionFooter, 1);
+    const previewActionFooter = creator.footerToolbar.actions.findIndex((action) => action.id === 'svd-preview');
+    if (previewActionFooter >= 0) creator.footerToolbar.actions.splice(previewActionFooter, 1);
+    const settingsActionFooter = creator.footerToolbar.actions.findIndex((action) => action.id === 'svd-settings');
+    if (settingsActionFooter >= 0) creator.footerToolbar.actions.splice(settingsActionFooter, 1);
+
+    creator.onElementAllowOperations.add((_, options) => {
+      // eslint-disable-next-line no-param-reassign
+      options.allowShowSettings = false;
     });
 
+    creator.onDefineElementMenuItems.add((_, options) => {
+      const settingsItemIndex = options.items.findIndex((option) => option.iconName === 'icon-settings_16x16');
+      if (settingsItemIndex >= 0) {
+        options.items.splice(settingsItemIndex, 1);
+      }
+    });
+
+    creatorRef.current = creator;
+  }
+
+  const creator = creatorRef.current;
+
+  // Only update these properties when necessary
+  useEffect(() => {
+    if (creator) {
+      creator.saveNo = form.getValues('saveNo');
+      creator.JSON = form.getValues('formula');
+      creator.locale = language;
+    }
+  }, [creator, form, language]);
+
+  const handleSaveSurvey = async () => {
+    if (!creator) return;
+    const formula = creator.JSON as SurveyFormula;
+    const saveNo = creator.saveNo || 0;
+    const success = await updateOrCreateSurvey({
+      ...form.getValues(),
+      formula,
+      saveNo,
+    });
     if (success) {
       void updateUsersSurveys();
       setIsOpenSaveSurveyDialog(false);
-
       toast.success(t('survey.editor.saveSurveySuccess'));
       navigate(`/${CREATED_SURVEYS_PAGE}`);
     }
   };
+
+  const pageBarsHeight = useElementHeight([FLOATING_BUTTONS_BAR_ID, FOOTER_ID]) + 10;
 
   const config: FloatingButtonsBarConfig = {
     buttons: [SaveButton(() => setIsOpenSaveSurveyDialog(true))],
@@ -124,18 +191,28 @@ const SurveyEditorPage = () => {
 
   return (
     <>
-      {isLoading ? <LoadingIndicatorDialog isOpen={isLoading} /> : null}
+      {isLoading && <LoadingIndicatorDialog isOpen={isLoading} />}
       {isFetching ? (
         <LoadingIndicatorDialog isOpen={isFetching} />
       ) : (
         <>
-          <SurveyEditor form={form} />
+          <div
+            className="survey-editor"
+            style={{ height: `calc(100% - ${pageBarsHeight}px)` }}
+          >
+            {creator && (
+              <SurveyCreatorComponent
+                creator={creator}
+                style={{ height: '100%', width: '100%' }}
+              />
+            )}
+          </div>
           <FloatingButtonsBar config={config} />
           <SaveSurveyDialog
             form={form}
             isOpenSaveSurveyDialog={isOpenSaveSurveyDialog}
             setIsOpenSaveSurveyDialog={setIsOpenSaveSurveyDialog}
-            submitSurvey={saveSurvey}
+            submitSurvey={handleSaveSurvey}
             isSubmitting={isLoading}
           />
         </>
